moved nsxt gateway service and constants into lib/nsxt
